style(csrf.ts): rename xsrf-token cookie name to XSRF-TOKEN
When following the guide in https://preview.adonisjs.com/guides/csrf-protection, it's understood
that when you enable `enableXsrfCookie`, AdonisJS will set a cookie named XSRF-TOKEN. Futhermore,
the template (shield.txt) says the following "When the following flag is enabled, AdoniJS will drop
`XSRF-TOKEN` cookie [...]". But all that doesn't happen, Adonis sets a cookie named `xsrf-token` and
a library like axios doesn't recognize it. So, this commit changes that.